requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==42.0.8
email-validator==2.2.0
gunicorn==22.0.0